Refactor immutable implementation to use a more generically named BuilderPojoPropertiesFactory. This factory should be reusable as is with other implementors of the builder pattern
