+++ core/src/main/java/org/jdbi/v3/core/mapper/reflect/internal/BuilderPojoPropertiesFactory.java
@@ -1,0 +1,28 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jdbi.v3.core.mapper.reflect.internal;
+
+import java.util.function.Supplier;
+
+import org.jdbi.v3.core.config.JdbiCache;
+import org.jdbi.v3.core.config.JdbiCaches;
+
+public interface BuilderPojoPropertiesFactory extends PojoPropertiesFactory {
+    JdbiCache<BuilderSpec<?, ?>, BuilderPojoProperties<?, ?>> BUILDER_CACHE =
+        JdbiCaches.declare(s -> s.type, BuilderPojoProperties::new);
+
+    static <T, B> PojoPropertiesFactory builder(Class<T> defn, Supplier<B> builder) {
+        return (t, config) -> BUILDER_CACHE.get(new BuilderSpec<>(t, config, defn, builder), config);
+    }
+}
